Adapter Pattern

+++ CodingPractice/Coding Main/LLD/DesignPattern/StructuralDesignPattern/Adapter Design Pattern/Adapter/IPoundAdapter.cs
@@ -2,6 +2,6 @@
 {
     public interface IPoundAdapter
     {
-        int GetWeightInPounds();
+        double GetWeightInPounds();
     }
 }
